requirements.txt
Flask
PyMuPDF
transformers
google.generativeai
firebase_admin
numpy